almost almost done

+++ philosophers.h
@@ -14,7 +14,7 @@ typedef	struct 	s_mutex{
 }				t_mutex;
 
 typedef	struct 	s_data{
-	u_int64_t	number_of_philosophers;
+	int			number_of_philosophers;
 	int			time_to_die;
 	int			time_to_eat;
 	int			time_to_sleep;
